Fix precise state issue
